task/BS-03-29: Add sorted devise feature specs
