+++ spec/features/homes/home_page_spec.rb
@@ -1,19 +1,16 @@
 RSpec.describe 'HomePage', type: :feature do
   let(:home_page) { Pages::HomePage::Home.new }
 
-  before do
-    create(:book)
-    home_page.load
-  end
+  before { home_page.load }
 
   it { expect(home_page).to have_current_path(root_path) }
 
   context 'with header elements' do
-    it { expect(home_page.header).to be_all_there }
-  end
-
-  context 'with slider elements' do
-    it { expect(home_page.slider).to have_buy_now_buttons(count: 1) }
+    it { expect(home_page.header).to have_brand_name }
+    it { expect(home_page.header).to have_home_link }
+    it { expect(home_page.header).to have_shop_link }
+    it { expect(home_page.header).to have_log_in_link }
+    it { expect(home_page.header).to have_sign_up_link }
   end
 
   context 'with get started elements' do
@@ -34,18 +31,5 @@ RSpec.describe 'HomePage', type: :feature do
     before { home_page.get_started.get_started_link.click }
 
     it { expect(home_page).to have_current_path(books_path) }
-  end
-
-  context 'when user signed in' do
-    let(:user) { create(:user) }
-    let(:sign_in_page) { Pages::SignIn.new }
-
-    before do
-      sign_in_page.load
-      sign_in_page.sign_in_form.authenticate_user user.email, user.password
-    end
-
-    it { expect(home_page.header).to have_my_account_link }
-    it { expect(home_page.header).to have_log_out_link }
   end
 end
