+++ spec/support/pages/devise/sign_up.rb
@@ -1,0 +1,11 @@
+module Pages
+  module Devise
+    class SignUp < SitePrism::Page
+      set_url '/users/sign_up'
+
+      section :auth_form_header, Sections::Devise::AuthFormHeader, '.container>div.text-center'
+      section :sign_up_form, Sections::Devise::SignUpForm, '.general-form'
+      sections :links, Sections::Devise::SignUpLinks, '.general-sign-text'
+    end
+  end
+end
