--- spec/support/sections/auth_form_header.rb
@@ -1,6 +1,0 @@
-module Sections
-  class AuthFormHeader < SitePrism::Section
-    element :page_name, '.general-login-title'
-    element :facebook_icon, '.general-login-icon'
-  end
-end
