+++ spec/support/sections/devise/auth_form_header.rb
@@ -1,0 +1,8 @@
+module Sections
+  module Devise
+    class AuthFormHeader < SitePrism::Section
+      element :page_name, '.general-login-title'
+      element :facebook_icon, '.general-login-icon'
+    end
+  end
+end
