+++ spec/support/sections/devise/sign_in_links.rb
@@ -1,0 +1,8 @@
+module Sections
+  module Devise
+    class SignInLinks < SitePrism::Section
+      element :sign_up_link, '.sign_up>a'
+      # element :confirm_account_link, '.confirm>a'
+    end
+  end
+end
