+++ spec/support/sections/devise/sign_up_links.rb
@@ -1,0 +1,8 @@
+module Sections
+  module Devise
+    class SignUpLinks < SitePrism::Section
+      element :log_in_link, '.log_in>a'
+      element :confirm_account_link, '.confirm>a'
+    end
+  end
+end
