--- spec/support/sections/sign_in_links.rb
@@ -1,6 +1,0 @@
-module Sections
-  class SignInLinks < SitePrism::Section
-    element :sign_up_link, '.sign_up>a'
-    # element :confirm_account_link, '.confirm>a'
-  end
-end
